refactor: remove unused GitHub maintainer detection code
- Delete 7 unused maintainer check methods from tool_policy.py (~450 lines)
- Remove GitHub API calls and cache file management
- Rename test_maintainer_bypass.py to test_development_source_bypass.py
- Clean up maintainer check mocks from remaining test files
