+++ src/ai_guardian/tool_policy.py
@@ -1499,457 +1499,6 @@ class ToolPolicyChecker:
 
         return "unknown"
 
-    def _get_git_repo_info(self) -> Optional[Tuple[str, str]]:
-        """
-        Extract owner/repo from git remote URL.
-
-        Returns:
-            tuple: (owner, repo) or None if not a git repo
-        """
-        try:
-            result = subprocess.run(
-                ["git", "config", "--get", "remote.origin.url"],
-                capture_output=True,
-                text=True,
-                timeout=5,
-                cwd=os.getcwd()
-            )
-            if result.returncode != 0:
-                return None
-
-            url = result.stdout.strip()
-            if not url:
-                return None
-
-            # Parse GitHub URL patterns:
-            # - https://github.com/owner/repo.git
-            # - git@github.com:owner/repo.git
-            # - https://github.com/owner/repo
-            if "github.com" not in url:
-                logger.info("❌ Maintainer bypass unavailable: Not a GitHub repository")
-                return None
-
-            # Extract owner/repo
-            if url.startswith("git@github.com:"):
-                path = url.replace("git@github.com:", "")
-            elif "github.com/" in url:
-                path = url.split("github.com/", 1)[1]
-            else:
-                return None
-
-            # Remove .git suffix
-            if path.endswith(".git"):
-                path = path[:-4]
-
-            # Split into owner/repo
-            parts = path.split("/")
-            if len(parts) >= 2:
-                owner, repo = parts[0], parts[1]
-                logger.debug(f"Detected git repo: {owner}/{repo}")
-                return owner, repo
-
-            return None
-
-        except Exception as e:
-            logger.debug(f"Failed to get git repo info: {e}")
-            return None
-
-    def _get_authenticated_github_user(self) -> Optional[str]:
-        """
-        Get GitHub username from authenticated gh CLI.
-
-        SECURITY: Only uses gh API which requires real OAuth token.
-        Does NOT trust environment variables or git config (easily spoofed).
-
-        Returns:
-            str: GitHub username or None if not authenticated
-        """
-        try:
-            result = subprocess.run(
-                ["gh", "api", "user", "--jq", ".login"],
-                capture_output=True,
-                text=True,
-                timeout=5
-            )
-            if result.returncode == 0:
-                username = result.stdout.strip()
-                if username:
-                    logger.debug(f"Authenticated GitHub user: {username}")
-                    return username
-            else:
-                logger.info("❌ Maintainer bypass unavailable: gh CLI not authenticated")
-                logger.info("   Fix: Run 'gh auth login' to authenticate")
-                return None
-        except FileNotFoundError:
-            logger.info("❌ Maintainer bypass unavailable: gh CLI not found")
-            logger.info("   Fix: Install gh CLI from https://cli.github.com/")
-            return None
-        except Exception as e:
-            logger.error(f"❌ Failed to get authenticated user: {e}")
-            return None
-
-        return None
-
-    def _check_github_collaborator(self, owner: str, repo: str, username: str) -> bool:
-        """
-        Check if user has write access to repo via GitHub API.
-
-        SECURITY: Requires real GitHub authentication, can't be spoofed.
-        Returns 204 if collaborator with write access, 404 if not.
-
-        Args:
-            owner: Repository owner
-            repo: Repository name
-            username: GitHub username to check
-
-        Returns:
-            bool: True if user has write access, False otherwise
-        """
-        max_attempts = 2
-        retry_delay = 1  # seconds
-
-        for attempt in range(1, max_attempts + 1):
-            try:
-                result = subprocess.run(
-                    ["gh", "api", f"repos/{owner}/{repo}/collaborators/{username}"],
-                    capture_output=True,
-                    timeout=10
-                )
-                is_collaborator = (result.returncode == 0)
-
-                if is_collaborator:
-                    logger.info(f"✅ Collaborator check: {username} on {owner}/{repo} = True")
-                else:
-                    logger.info(f"❌ Collaborator check: {username} on {owner}/{repo} = False")
-
-                return is_collaborator
-
-            except subprocess.TimeoutExpired as e:
-                if attempt < max_attempts:
-                    logger.info(f"GitHub API timeout (attempt {attempt}/{max_attempts}), retrying in {retry_delay}s...")
-                    time.sleep(retry_delay)
-                else:
-                    logger.error(f"❌ GitHub API timeout after {max_attempts} attempts: {e}")
-                    return False
-
-            except Exception as e:
-                if attempt < max_attempts:
-                    logger.info(f"GitHub API error (attempt {attempt}/{max_attempts}): {e}, retrying in {retry_delay}s...")
-                    time.sleep(retry_delay)
-                else:
-                    logger.error(f"❌ Collaborator check failed after {max_attempts} attempts: {e}")
-                    return False
-
-        return False
-
-    def _get_maintainer_cache(self) -> Optional[bool]:
-        """
-        Read maintainer status from cache file.
-
-        Returns:
-            bool or None: Cached maintainer status or None if not cached/expired
-        """
-        try:
-            cache_dir = Path.home() / ".cache" / "ai-guardian"
-            cache_file = cache_dir / "maintainer-status.json"
-
-            if not cache_file.exists():
-                return None
-
-            # Read and parse cache file
-            try:
-                with open(cache_file, 'r') as f:
-                    cache_data = json.load(f)
-            except json.JSONDecodeError as e:
-                logger.error(f"❌ Corrupt cache file: {cache_file}")
-                logger.error(f"   JSON decode error: {e}")
-                logger.error(f"   Fix: Delete corrupt cache with: rm {cache_file}")
-                return None
-            except OSError as e:
-                logger.error(f"❌ Failed to read cache file: {e}")
-                return None
-
-            # Get current repo info
-            repo_info = self._get_git_repo_info()
-            if not repo_info:
-                return None
-
-            owner, repo = repo_info
-            repo_key = f"{owner}/{repo}"
-
-            # Check if we have a cache entry for this repo
-            repositories = cache_data.get("repositories", {})
-            if not isinstance(repositories, dict):
-                logger.warning(f"Cache 'repositories' field is not a dict: {type(repositories)}")
-                return None
-
-            repo_cache = repositories.get(repo_key)
-            if not repo_cache:
-                return None
-
-            if not isinstance(repo_cache, dict):
-                logger.warning(f"Cache entry for {repo_key} is not a dict: {type(repo_cache)}")
-                return None
-
-            # Validate cache structure
-            checked_at = repo_cache.get("checked_at")
-            if not checked_at:
-                logger.warning(f"Cache entry missing 'checked_at' field for {repo_key}")
-                return None
-
-            if not isinstance(checked_at, str):
-                logger.error(f"Invalid 'checked_at' type: {type(checked_at)}, expected str")
-                return None
-
-            ttl_hours = cache_data.get("ttl_hours", 24)
-
-            # Parse timestamp and check expiry
-            try:
-                checked_time = datetime.fromisoformat(checked_at.replace('Z', '+00:00'))
-                current_time = datetime.now(timezone.utc)
-                age = current_time - checked_time
-
-                # Check if still within TTL
-                if age < timedelta(hours=ttl_hours):
-                    is_maintainer = repo_cache.get("is_maintainer", False)
-                    age_str = f"{int(age.total_seconds() // 3600)}h {int((age.total_seconds() % 3600) // 60)}m"
-                    logger.info(f"✅ Using cached maintainer status: {is_maintainer} (age: {age_str})")
-                    return is_maintainer
-                else:
-                    logger.debug(f"Cache expired for {repo_key} (age: {age}, TTL: {ttl_hours}h)")
-                    return None
-
-            except ValueError as e:
-                logger.error(f"❌ Invalid timestamp format in cache: '{checked_at}'")
-                logger.error(f"   Parse error: {e}")
-                logger.error(f"   Fix: Delete corrupt cache with: rm {cache_file}")
-                return None
-
-        except Exception as e:
-            # Catch-all for unexpected errors - log with full traceback
-            logger.error(f"❌ Unexpected error reading maintainer cache: {e}")
-            import traceback
-            logger.error(traceback.format_exc())
-            return None
-
-    def _cache_maintainer_status(self, is_maintainer: bool) -> None:
-        """
-        Write maintainer status to cache file.
-
-        Args:
-            is_maintainer: Whether user is a maintainer
-        """
-        try:
-            repo_info = self._get_git_repo_info()
-            if not repo_info:
-                return
-
-            owner, repo = repo_info
-            repo_key = f"{owner}/{repo}"
-
-            # Get authenticated user
-            username = self._get_authenticated_github_user()
-            if not username:
-                return
-
-            cache_dir = Path.home() / ".cache" / "ai-guardian"
-            cache_file = cache_dir / "maintainer-status.json"
-
-            # Create cache directory if needed
-            cache_dir.mkdir(parents=True, exist_ok=True)
-
-            # Load existing cache or create new
-            cache_data = {}
-            if cache_file.exists():
-                try:
-                    with open(cache_file, 'r') as f:
-                        cache_data = json.load(f)
-                except:
-                    cache_data = {}
-
-            # Update cache data
-            if "repositories" not in cache_data:
-                cache_data["repositories"] = {}
-
-            cache_data["version"] = 1
-            # Get TTL from env var with upper bound (max 168 hours = 7 days)
-            ttl_hours = int(os.environ.get("AI_GUARDIAN_MAINTAINER_CACHE_TTL_HOURS", "24"))
-            cache_data["ttl_hours"] = min(ttl_hours, 168)  # Cap at 7 days
-            cache_data["repositories"][repo_key] = {
-                "username": username,
-                "is_maintainer": is_maintainer,
-                "checked_at": datetime.now(timezone.utc).isoformat()
-            }
-
-            # Write cache file
-            with open(cache_file, 'w') as f:
-                json.dump(cache_data, f, indent=2)
-
-            logger.debug(f"Cached maintainer status for {repo_key}: {is_maintainer}")
-
-        except Exception as e:
-            logger.debug(f"Failed to cache maintainer status: {e}")
-
-    def _is_github_maintainer_cached(self) -> bool:
-        """
-        Check if user is a maintainer of the current repository (with cache).
-
-        Returns:
-            bool: True if user is a maintainer, False otherwise
-        """
-        # Check cache first
-        cached = self._get_maintainer_cache()
-        if cached is not None:
-            return cached
-
-        try:
-            # Get repository info
-            repo_info = self._get_git_repo_info()
-            if not repo_info:
-                logger.info("❌ Maintainer bypass unavailable: Not a git repository")
-                return False
-
-            owner, repo = repo_info
-
-            # Get authenticated GitHub user
-            username = self._get_authenticated_github_user()
-            if not username:
-                # Error already logged in _get_authenticated_github_user
-                return False
-
-            # Check collaborator status
-            is_maintainer = self._check_github_collaborator(owner, repo, username)
-
-            # Cache result
-            self._cache_maintainer_status(is_maintainer)
-
-            return is_maintainer
-
-        except Exception as e:
-            logger.error(f"❌ Maintainer check failed with unexpected error: {e}")
-            import traceback
-            logger.error(traceback.format_exc())
-            return False
-
-    def _diagnose_maintainer_bypass(self) -> str:
-        """
-        Diagnose why maintainer bypass is not working.
-
-        Returns:
-            str: User-friendly diagnostic message with actionable steps
-        """
-        # Check if it's a GitHub repo
-        repo_info = self._get_git_repo_info()
-        if not repo_info:
-            return (
-                "❌ Not a GitHub repository\n"
-                "   Maintainer bypass only works for GitHub repositories.\n"
-                "   \n"
-                "   Current directory must be a git repository with a GitHub remote.\n"
-            )
-
-        owner, repo = repo_info
-        repo_url = f"github.com/{owner}/{repo}"
-
-        # Check gh CLI authentication
-        try:
-            result = subprocess.run(
-                ["gh", "auth", "status"],
-                capture_output=True,
-                text=True,
-                timeout=5
-            )
-            if result.returncode != 0:
-                return (
-                    f"❌ GitHub CLI not authenticated\n"
-                    f"   Repository: {repo_url}\n"
-                    f"   \n"
-                    f"   Fix: Run 'gh auth login' to authenticate with GitHub\n"
-                )
-        except FileNotFoundError:
-            return (
-                f"❌ GitHub CLI not installed\n"
-                f"   Repository: {repo_url}\n"
-                f"   \n"
-                f"   Fix: Install gh CLI from https://cli.github.com/\n"
-            )
-
-        # Get authenticated user
-        username = self._get_authenticated_github_user()
-        if not username:
-            return (
-                f"❌ Could not determine GitHub username\n"
-                f"   Repository: {repo_url}\n"
-                f"   \n"
-                f"   Fix: Run 'gh auth login' and ensure authentication succeeds\n"
-            )
-
-        # Check collaborator status
-        try:
-            result = subprocess.run(
-                ["gh", "api", f"repos/{owner}/{repo}/collaborators/{username}"],
-                capture_output=True,
-                timeout=10
-            )
-            if result.returncode != 0:
-                return (
-                    f"❌ Not a repository maintainer\n"
-                    f"   Repository: {repo_url}\n"
-                    f"   GitHub User: {username}\n"
-                    f"   \n"
-                    f"   Maintainer bypass requires write access to the repository.\n"
-                    f"   Only repository owners and collaborators can edit source files.\n"
-                    f"   \n"
-                    f"   To check your access: gh api repos/{owner}/{repo}/collaborators/{username}\n"
-                )
-        except subprocess.TimeoutExpired:
-            return (
-                f"❌ GitHub API timeout\n"
-                f"   Repository: {repo_url}\n"
-                f"   \n"
-                f"   Could not verify collaborator status due to timeout.\n"
-                f"   Check your network connection and try again.\n"
-            )
-        except Exception as e:
-            return (
-                f"❌ GitHub API error\n"
-                f"   Repository: {repo_url}\n"
-                f"   Error: {e}\n"
-                f"   \n"
-                f"   Could not verify collaborator status.\n"
-            )
-
-        # Check cache status
-        cache_file = Path.home() / ".cache" / "ai-guardian" / "maintainer-status.json"
-        cache_status = "No cache found"
-        if cache_file.exists():
-            try:
-                with open(cache_file, 'r') as f:
-                    cache_data = json.load(f)
-                    repo_key = f"{owner}/{repo}"
-                    if repo_key in cache_data.get("repositories", {}):
-                        repo_cache = cache_data["repositories"][repo_key]
-                        checked_at = repo_cache.get("checked_at", "unknown")
-                        is_maintainer = repo_cache.get("is_maintainer", False)
-                        cache_status = f"Cached: is_maintainer={is_maintainer}, checked_at={checked_at}"
-            except:
-                cache_status = "Cache file corrupt"
-
-        # All checks passed but still blocked - something went wrong
-        return (
-            f"⚠️  Unexpected: All checks passed but bypass failed\n"
-            f"   Repository: {repo_url}\n"
-            f"   GitHub User: {username}\n"
-            f"   Collaborator Status: ✅ Write access confirmed\n"
-            f"   Cache: {cache_status}\n"
-            f"   \n"
-            f"   This may be a bug in ai-guardian. Try:\n"
-            f"   1. Clear cache: rm {cache_file}\n"
-            f"   2. Try the operation again\n"
-            f"   3. Check logs for more details\n"
-            f"   4. Report issue: https://github.com/itdove/ai-guardian/issues\n"
-        )
-
     def _load_config(self) -> Dict:
         """
         Load and merge tool policy configurations.
